# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Code Red Backend API
  version: 1.3.0
  description: >
    Production-style backend contract for the Code Red coaching institute platform.
    Canonical routes are versioned under `/api/v1`. This revision aligns the contract
    with the current frontend implementation, including authority-based navigation,
    inventory requests, payroll views, promotion flows, and faculty material publishing.
    Access control follows the frontend role families: student, faculty, parent,
    and administration (`admin`, `administration`, `director`, `superadmin`). Request
    payloads are validated with Pydantic, some background work is dispatched through
    Celery using Redis, and throttled endpoints may return HTTP 429.
servers:
  - url: http://localhost:3500
    description: Local Flask API server
  - url: http://localhost:3000
    description: Frontend dev server proxy, if configured
...
paths:
...
  /api/v1/students:
    get:
      tags: [Students]
      summary: List students
      description: Accessible to faculty and administration.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: class
          schema: { type: string }
        - in: query
          name: section
          schema: { type: string }
      responses:
        '200':
          description: Student list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
        '403':
          $ref: '#/components/responses/Error403'
  /api/v1/students/{studentId}:
    get:
      tags: [Students]
      summary: Get student by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StudentIdPath'
      responses:
        '200':
          description: Student details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '404':
          $ref: '#/components/responses/Error404'
  /api/v1/students/me:
    get:
      tags: [Students]
      summary: Get current student profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
  /api/v1/students/me/subjects:
    get:
      tags: [Students]
      summary: Get current student subjects
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Subject list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subject'
  /api/v1/students/me/subjects/{subjectId}/content:
    get:
      tags: [Students]
      summary: Get current student subject content tree
      description: >
        Returns the week-grouped subject content used by the student subject details page,
        including weekly contents, study materials, and assessments.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SubjectIdPath'
      responses:
        '200':
          description: Subject content tree
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentSubjectContent'
        '404':
          $ref: '#/components/responses/Error404'
  /api/v1/students/me/subjects/{subjectId}/chat:
    post:
      tags: [Students, AI]
      summary: Ask a grounded chatbot question about a student subject
      description: >
        Uses uploaded subject materials plus stored AI runtime settings to answer a student's
        question with cited source snippets. Falls back to grounded extractive answers when no
        external provider is configured or available.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SubjectIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentSubjectChatRequest'
      responses:
        '200':
          description: Grounded subject answer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentSubjectChatResponse'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
  /api/v1/students/me/performance:
    get:
      tags: [Students]
      summary: Get current student performance summary
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Performance summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerformanceSummary'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  parameters:
    StudentIdPath:
      in: path
      name: studentId
      required: true
      schema: { type: string }
...
    SubjectIdPath:
      in: path
      name: subjectId
      required: true
      schema: { type: string }
...
  responses:
...
    Error403:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Error404:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    Error422:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Error429:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: object
              additionalProperties: true
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        firstName: { type: string }
        lastName: { type: string }
        role: { type: string }
        employeeCode: { type: string, nullable: true }
        roles:
          type: array
          items: { type: string }
        roleScope: { type: string }
        title: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        profilePicture: { type: string, nullable: true }
    Student:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            class: { type: string, nullable: true }
            section: { type: string, nullable: true }
            rollNumber: { type: string, nullable: true }
            enrollmentNo: { type: string, nullable: true }
            classId: { type: string, nullable: true }
...
    Subject:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        code: { type: string, nullable: true }
        teacherId: { type: string, nullable: true }
        facultyId: { type: string, nullable: true }
        credits: { type: integer, nullable: true }
...
    Material:
      type: object
      properties:
        id: { type: string }
        subjectId: { type: string, nullable: true }
        subjectName: { type: string, nullable: true }
        chapterId: { type: string, nullable: true }
        chapterTitle: { type: string, nullable: true }
        title: { type: string }
        unit: { type: string, nullable: true }
        week: { type: string, nullable: true }
        type: { type: string }
        description: { type: string, nullable: true }
        fileName: { type: string, nullable: true }
        documentId: { type: string, nullable: true }
        documentName: { type: string, nullable: true }
        documentUrl: { type: string, nullable: true }
        externalUrl: { type: string, nullable: true }
        uploadedAt: { type: string, format: date-time, nullable: true }
        storagePath: { type: string, nullable: true }
        uploadedBy: { type: string, nullable: true }
        isFacultyUpload: { type: boolean, nullable: true }
    StudentSubjectChatHistoryMessage:
      type: object
      properties:
        role: { type: string }
        content: { type: string }
      required: [role, content]
    StudentSubjectChatRequest:
      type: object
      properties:
        question: { type: string }
        history:
          type: array
          items:
            $ref: '#/components/schemas/StudentSubjectChatHistoryMessage'
      required: [question]
    StudentSubjectChatCitation:
      type: object
      properties:
        materialId: { type: string }
        materialTitle: { type: string }
        snippet: { type: string }
    StudentSubjectChatResponse:
      type: object
      properties:
        subjectId: { type: string }
        subjectName: { type: string }
        question: { type: string }
        answer: { type: string }
        confidence: { type: string }
        citations:
          type: array
          items:
            $ref: '#/components/schemas/StudentSubjectChatCitation'
        followUpQuestions:
          type: array
          items: { type: string }
    SubjectWeekContent:
      type: object
      properties:
        id: { type: string }
        weekLabel: { type: string }
        contents:
          type: array
          items:
            type: object
            additionalProperties: true
        materials:
          type: array
          items:
            $ref: '#/components/schemas/Material'
        assignments:
          type: array
          items:
            $ref: '#/components/schemas/Assignment'
    StudentSubjectContent:
      type: object
      properties:
        subject:
          $ref: '#/components/schemas/Subject'
        weeks:
          type: array
          items:
            $ref: '#/components/schemas/SubjectWeekContent'
...
    Assignment:
      type: object
      properties:
        id: { type: string }
        subjectId: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        dueDate: { type: string }
        totalMarks: { type: integer }
        status: { type: string }
    PerformanceSummary:
      type: object
      properties:
        average: { type: number }
        rank: { type: integer }
        totalStudents: { type: integer }
...